Update: We can now download songs from Text File in Order
User can now place their songs in the `yt_urls.txt` file in the order
that they would like to download and that audio function will prefix the
value of `1. `, `2. ` with each URL that is in the text file.

Hence, at the end, user will get "numbered" song names in the
`downloaded_audio_ordered` directory / folder.

INFO: I will update the `learning.md` file later on.